requirements.txt
# Core dependencies
simple-salesforce
python-dotenv

# Utilities
typing-extensions